Revert "pysdk: add command status (#5856)"
This reverts commit 2a5a969c94161f5404deef49337145d139017822.
